auth_service/requirements.txt
fastapi[standard]
fastapi-cors
uvicorn[standard]
pytest
pydantic
passlib[bcrypt]

psycopg2-binary
alembic
sqlalchemy

requests
oauthlib
authlib
pyOpenSSL
python-jose
asyncpg

aio_pika
pyjwt
greenlet
python-dotenv
requests

itsdangerous